Moved Auth-Code to Core
