+++ application/Core/Auth/Adapter.php
@@ -7,27 +7,32 @@
  * To change this template use File | Settings | File Templates.
  */
 
-namespace CoreApi\Service\Auth;
+namespace Core\Auth;
 
-class Bypass
+class Adapter
     implements \Zend_Auth_Adapter_Interface
 {
 
     const NOT_FOUND_MESSAGE 	= 'Unknown login!';
     const CREDINTIALS_MESSAGE 	= 'Wrong Password!';
     const NOT_ACTIVATED_MESSAGE = 'Account is not yet activated!';
-    const UNKNOWN_FAILURE 	= 'Unknown error!';
+    const UNKNOWN_FAILURE 		= 'Unknown error!';
 
-    private $em;
 
+    /** @var \Core\Entity\User $user */
     private $user;
+    
+    /** @var \Core\Entity\Login $login */
+    private $login;
+    
+    /** @var string $password */
+    private $password;
 
 
-    public function __construct($user)
+    public function __construct(\Core\Entity\Login $login, $password)
     {
-        $this->em = \Zend_Registry::get('doctrine')->getEntityManager();
-
-        $this->user = $user;
+        $this->login = $login;
+        $this->password = $password;
     }
 
     /**
@@ -38,24 +43,36 @@ class Bypass
      */
     public function authenticate()
     {
-		$user = $this->user;
-  
         // User Not Found:
-        if(is_null($user))
+        if(is_null($this->login))
         {
             return $this->authResult(
                 \Zend_Auth_Result::FAILURE_IDENTITY_NOT_FOUND,
                 self::NOT_FOUND_MESSAGE
             );
         }
+        
+        
+        /** @var $user \Entity\User */
+        $this->user = $this->login->getUser();
 
 
         // User Not Activated:
-        if($user->getState() != \Core\Entity\User::STATE_ACTIVATED || is_null($user->getLogin()))
+        if($this->user->getState() != \Core\Entity\User::STATE_ACTIVATED)
         {
             return $this->authResult(
                 \Zend_Auth_Result::FAILURE_IDENTITY_AMBIGUOUS,
                 self::NOT_ACTIVATED_MESSAGE
+            );
+        }
+
+
+        // User with wrong Password:
+        if(!$this->login->checkPassword($this->password))
+        {
+            return $this->authResult(
+                \Zend_Auth_Result::FAILURE_CREDENTIAL_INVALID,
+                self::CREDINTIALS_MESSAGE
             );
         }
 
@@ -65,17 +82,17 @@ class Bypass
 
 
      /**
-      * Factory for Zend_Auth_Result
-      *
-      * @param integer    The Result code, see Zend_Auth_Result
-      * @param mixed      The Message, can be a string or array
-      * @return Zend_Auth_Result
-      */
+     * Factory for Zend_Auth_Result
+     *
+     *@param integer    The Result code, see Zend_Auth_Result
+     *@param mixed      The Message, can be a string or array
+     *@return Zend_Auth_Result
+     */
     private function authResult($code, $messages = array())
 	{
         if( !is_array( $messages ) )
         {	$messages = array($messages);	}
-        
+
 		return new \Zend_Auth_Result($code, $this->user->getId(), $messages);
     }
 }
